fix : prompt_new_v2를 기본 라우터로 수정
