+++ src/components/lnb/HomeLnb.tsx
@@ -1,14 +1,13 @@
 "use client";
 
-import React from "react";
-import LNB, { MenuItemsType } from "../lnb";
-import { useDeviceSize } from "@components/DeviceContext";
-import Text from "@/components/common/Text/Text";
 import Icon from "@/components/common/Icon";
+import Text from "@/components/common/Text/Text";
 import useToast from "@/hooks/useToast";
-import Button from "../common/Button/Button";
-import Link from "next/link";
 import { useUser } from "@/hooks/useUser";
+import { useDeviceSize } from "@components/DeviceContext";
+import Link from "next/link";
+import Button from "../common/Button/Button";
+import LNB, { MenuItemsType } from "../lnb";
 
 interface HomeLnbType {
     initialMenu: string;
@@ -73,7 +72,7 @@ const HomeLnb = ({ initialMenu }: HomeLnbType) => {
     };
 
     const newPropmptButton = userData.isLogin ? (
-        <Link href="/prompt-new-v2">
+        <Link href="/prompt-new">
             <Button
                 id="register-prompt"
                 style={{ padding: "8px 12px", gap: 2 }}
