+++ src/components/promptNew/PreviewSection.tsx
@@ -1,15 +1,17 @@
+"use client";
+
 import Text from "@/components/common/Text/Text";
-import ExampleBox from "./Example/ExampleBox";
-import ExampleContent from "./Example/ExampleContent";
 import { PromptSchemaType } from "@/schema/PromptSchema";
 import { extractOptions } from "@/utils/promptUtils";
-import { Flex } from "antd";
+import { useDeviceSize } from "@components/DeviceContext";
 import { useMemo } from "react";
 import { useFormContext } from "react-hook-form";
 import styled from "styled-components";
+import ExampleContent from "./Example/ExampleContent";
 
 function PreviewSection() {
     const { watch } = useFormContext<PromptSchemaType>();
+    const { isUnderTablet, isMobile } = useDeviceSize();
     const title = watch("title");
     const description = watch("description");
     const prompt_template = watch("prompt_template");
@@ -18,42 +20,24 @@ function PreviewSection() {
     }, [prompt_template]);
 
     return (
-        <Box>
+        <Box $isUnderTablet={isUnderTablet}>
             <Text font="h2_20_bold">내 프롬프트 미리보기</Text>
 
-            <Flex vertical gap={30} style={{ marginTop: "24px" }}>
-                <Flex vertical gap={6}>
-                    <ExampleBox
-                        defaultValue="프롬프트 제목이 이곳에 표시됩니다."
-                        value={title}
-                        font="b1_18_semi"
-                        color="G_700"
-                    />
-                    <ExampleBox
-                        defaultValue="프롬프트 설명이 이곳에 표시됩니다."
-                        value={description}
-                        font="b3_14_reg"
-                        color="G_500"
-                    />
-                </Flex>
-
-                <ExampleContent
-                    defaultValue="프롬프트 내용에 따른 미리보기가 이곳에 표시됩니다."
-                    value={inputs}
-                />
-            </Flex>
+            <ExampleContent
+                defaultValue="프롬프트 내용에 따른 미리보기가 이곳에 표시됩니다."
+                value={inputs}
+            />
         </Box>
     );
 }
 
 export default PreviewSection;
 
-const Box = styled.div`
-    flex: 3;
-    max-width: 30%;
-
+const Box = styled.div<{ $isUnderTablet: boolean }>`
+    /* max-width: 100%; */
     border-radius: 16px;
     border: 1.5px solid ${({ theme }) => theme.colors.G_100};
     background: #fff;
     padding: 20px;
+    width: ${({ $isUnderTablet }) => ($isUnderTablet ? "97%" : "30%")};
 `;
